# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "APIKeyErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/APIKeyErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "APIKeyErrorResponse",
        "type": "object"
      },
...
      "AuthErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/AuthErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "AuthErrorResponse",
        "type": "object"
      },
...
      "BillingCardSetupResponse": {
        "properties": {
          "client_secret": {
            "title": "Client Secret",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "client_secret"
        ],
        "title": "BillingCardSetupResponse",
        "type": "object"
      },
      "BillingCardSetupStatus": {
        "properties": {
          "registered": {
            "title": "Registered",
            "type": "boolean"
          }
        },
        "required": [
          "registered"
        ],
        "title": "BillingCardSetupStatus",
        "type": "object"
      },
...
      "BillingCheckoutForm": {
        "additionalProperties": false,
        "properties": {
          "currency": {
            "enum": [
              "krw",
              "usd"
            ],
            "title": "Currency",
            "type": "string"
          },
          "plan": {
            "enum": [
              "monthly",
              "annual",
              "pro_monthly",
              "pro_annual"
            ],
            "title": "Plan",
            "type": "string"
          },
          "request_id": {
            "description": "Reuse this UUID when retrying the same setup request.",
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "plan",
          "currency"
        ],
        "title": "BillingCheckoutForm",
        "type": "object"
      },
      "BillingCheckoutStatusResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "paid": {
            "title": "Paid",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "open",
              "complete",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "paid"
        ],
        "title": "BillingCheckoutStatusResponse",
        "type": "object"
      },
...
      "BillingErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/BillingErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "BillingErrorResponse",
        "type": "object"
      },
...
      "BillingSetupForm": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "description": "Reuse this UUID when retrying the same setup request.",
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "BillingSetupForm",
        "type": "object"
      },
      "BillingSetupResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url"
        ],
        "title": "BillingSetupResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InternalErrorResponse": {
        "properties": {
          "error": {
            "const": "INTERNAL_ERROR",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "InternalErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "B4 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/billing/card-setups": {
      "post": {
        "operationId": "create_card_setup_api_v1_billing_card_setups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingSetupForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingCardSetupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_INVALID",
                    "message": "Invalid billing webhook."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 401 Create Card Setup Api V1 Billing Card Setups Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 403 Create Card Setup Api V1 Billing Card Setups Post"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BillingErrorResponse"
                    }
                  ],
                  "title": "Response 404 Create Card Setup Api V1 Billing Card Setups Post"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_CHECKOUT_CONFLICT",
                    "message": "An existing subscription or another checkout requires attention."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "502": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_UNAVAILABLE",
                    "message": "Payment provider is temporarily unavailable."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_UNAVAILABLE",
                    "message": "Billing webhook is not configured."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Card Setup",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/card-setups/{intent_id}": {
      "get": {
        "operationId": "card_setup_status_api_v1_billing_card_setups__intent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "pattern": "^seti_[A-Za-z0-9]+$",
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingCardSetupStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_INVALID",
                    "message": "Invalid billing webhook."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 401 Card Setup Status Api V1 Billing Card Setups  Intent Id  Get"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 403 Card Setup Status Api V1 Billing Card Setups  Intent Id  Get"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BillingErrorResponse"
                    }
                  ],
                  "title": "Response 404 Card Setup Status Api V1 Billing Card Setups  Intent Id  Get"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_CHECKOUT_CONFLICT",
                    "message": "An existing subscription or another checkout requires attention."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "502": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_UNAVAILABLE",
                    "message": "Payment provider is temporarily unavailable."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_UNAVAILABLE",
                    "message": "Billing webhook is not configured."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Card Setup Status",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/checkout-sessions": {
      "post": {
        "operationId": "create_checkout_api_v1_billing_checkout_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingCheckoutForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingSetupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_INVALID",
                    "message": "Invalid billing webhook."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 401 Create Checkout Api V1 Billing Checkout Sessions Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 403 Create Checkout Api V1 Billing Checkout Sessions Post"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BillingErrorResponse"
                    }
                  ],
                  "title": "Response 404 Create Checkout Api V1 Billing Checkout Sessions Post"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_CHECKOUT_CONFLICT",
                    "message": "An existing subscription or another checkout requires attention."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "502": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_UNAVAILABLE",
                    "message": "Payment provider is temporarily unavailable."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_UNAVAILABLE",
                    "message": "Billing webhook is not configured."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Checkout",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/checkout-sessions/{session_id}": {
      "get": {
        "operationId": "checkout_status_api_v1_billing_checkout_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "pattern": "^cs_[A-Za-z0-9_]+$",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingCheckoutStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_INVALID",
                    "message": "Invalid billing webhook."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 401 Checkout Status Api V1 Billing Checkout Sessions  Session Id  Get"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 403 Checkout Status Api V1 Billing Checkout Sessions  Session Id  Get"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BillingErrorResponse"
                    }
                  ],
                  "title": "Response 404 Checkout Status Api V1 Billing Checkout Sessions  Session Id  Get"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_CHECKOUT_CONFLICT",
                    "message": "An existing subscription or another checkout requires attention."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "502": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_UNAVAILABLE",
                    "message": "Payment provider is temporarily unavailable."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_UNAVAILABLE",
                    "message": "Billing webhook is not configured."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Checkout Status",
        "tags": [
          "Billing"
        ]
      }
    },
...
    "/config": {
      "get": {
        "operationId": "config_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Config"
      }
    },
...
  }
}